+++ ProjectB/Logic/PaymentLogic/IPaymentLogic.cs
@@ -1,4 +1,4 @@
 public interface IPaymentLogic
 {
-    void Payment(decimal price);
+    void Payment(double price);
 }
