+++ ProjectB/Logic/PaymentLogic/PaypalPaymentLogic.cs
@@ -1,6 +1,6 @@
 public class PaypalPaymentLogic : IPaymentLogic
 {
-    public void Payment(decimal price)
+    public void Payment(double price)
     {
         // Als er Paypal word gekozen e-mail en wachtwoord vragen.  
         // Hiervoor hoeven geen checks gedaan worden, aangezien de e-mail en wachtwoord bij PayPal zijn aangemaakt.
